refactor: move fixture to conftest.
